refactor(bridge): split inbound message handling into per-method modules
The inbound surface (downstream → editor) was concentrated in
actor/reader.rs, so the structure gave no hint of which server messages
the bridge supports. The outbound textDocument/* side already uses one
file per method as a discoverable map; this extends that axis to the
inbound side.

Each LSP namespace becomes a directory of per-method files — client/,
window/, workspace/, plus the inbound scratch publishDiagnostics discard
under text_document/. reader.rs keeps only the shared transport and a
dispatcher that routes each method to its module. Pure structural Tidy:
behavior and log targets are unchanged.

$/progress sits under window/ alongside window/workDoneProgress/create
despite being $-namespaced: the two are one feature sharing the same
ProgressRegistry token remapping, so they stay cohesive. The mutable
WorkspaceFolderSet data type moves to workspace/folder_set.rs to avoid
colliding with the workspace/workspaceFolders pull handler.

Refs #393
